infra/requirements.txt
# Infra/SRE Bot Requirements
# Install with: pip install -r infra/requirements.txt

# CLI Framework
click>=8.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.24.0

# Pretty Output
rich>=13.0.0

# Provider SDKs
stripe>=6.0.0
supabase>=2.0.0
psycopg2-binary>=2.9.0
PyGithub>=1.59.0

# Optional: for future async support
# asyncio already in stdlib

